requirements.txt
python-telegram-bot==21.10
python-dotenv==1.1.0
google-generativeai==0.8.5
